Move unit tests to a tests subdirectory and begin splitting it out into files
