+++ tests/mocks.py
@@ -1,0 +1,35 @@
+import datetime
+import json
+from unittest.mock import Mock, MagicMock, patch, call
+
+class Session(object):
+    """
+    Python reqeusts.Session mockery class
+    """
+    request = None
+    headers = None
+
+class Response(object):
+    """Mock class for emulating requests.Response objects
+
+    Look for existing use of this class for examples on how to use.
+    """
+    def __init__(self, code, text, method='GET', url=None):
+        super(Response, self).__init__()
+        self.status_code = code
+        self.text = text
+        self.ok = code < 400
+        self.headers = MagicMock()
+        if url:
+            self.url = url
+        else:
+            self.url = 'https://api.pagerduty.com'
+        self.elapsed = datetime.timedelta(0,1.5)
+        self.request = Mock(url=self.url)
+        self.headers = {'date': 'somedate',
+            'x-request-id': 'xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx'}
+        self.request.method = method
+        self.json = MagicMock()
+        self.json.return_value = json.loads(text)
+
+
